+++ firebase-appdistribution-api/CHANGELOG.md
@@ -1,4 +1,19 @@
 # Unreleased
+* [changed] Added Kotlin extensions (KTX) APIs from 
+  `com.google.firebase:firebase-appdistribution-api-ktx`
+  to `com.google.firebase:firebase-appdistribution-api` under the 
+  `com.google.firebase.appdistribution` package.
+  For details, see the
+  [FAQ about this initiative](https://firebase.google.com/docs/android/kotlin-migration)
+
+* [deprecated] All the APIs from `com.google.firebase:firebase-appdistribution-api-ktx` have been 
+  added to
+  `com.google.firebase:firebase-appdistribution-api` under the 
+  `com.google.firebase.appdistribution` package,
+  and all the Kotlin extensions (KTX) APIs in `com.google.firebase:firebase-appdistribution-api-ktx`
+  are now deprecated. As early as April 2024, we'll no longer release KTX modules. For details, 
+  see the
+  [FAQ about this initiative](https://firebase.google.com/docs/android/kotlin-migration)
 
 # 16.0.0-beta09
 * [feature] Improved development mode to allow all API calls to be made without having to sign in.