+++ firebase-inappmessaging/CHANGELOG.md
@@ -1,4 +1,14 @@
 # Unreleased
+* [changed] Added Kotlin extensions (KTX) APIs from `com.google.firebase:firebase-inappmessaging-ktx`
+  to `com.google.firebase:firebase-inappmessaging` under the `com.google.firebase.inappmessaging` package.
+  For details, see the
+  [FAQ about this initiative](https://firebase.google.com/docs/android/kotlin-migration)
+
+* [deprecated] All the APIs from `com.google.firebase:firebase-inappmessaging-ktx` have been added to
+  `com.google.firebase:firebase-inappmessaging` under the `com.google.firebase.inappmessaging` package,
+  and all the Kotlin extensions (KTX) APIs in `com.google.firebase:firebase-inappmessaging-ktx` are
+  now deprecated. As early as April 2024, we'll no longer release KTX modules. For details, see the
+  [FAQ about this initiative](https://firebase.google.com/docs/android/kotlin-migration)
 
 
 # 20.3.5